pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"

[project]
name = "workflow-recorder"
version = "0.1.0"
description = "Windows daemon that captures screenshots and analyzes employee workflows via GPT vision"
requires-python = ">=3.9"
dependencies = [
    "mss>=9.0",
    "Pillow>=10.0",
    "openai>=1.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "structlog>=24.0",
    "imagehash>=4.3",
    "psutil>=5.9",
    "pywin32>=306; sys_platform == 'win32'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
workflow-recorder = "workflow_recorder.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/workflow_recorder"]
